Silly bug introduced while debugging it
The cli stuff must have fixed it.

+++ RF69.h
@@ -15,7 +15,7 @@ namespace RF69 {
     extern uint16_t fifooverrun; 
     extern uint16_t busyCount;
     extern uint16_t underrun;
-    extern uint8_t interruptTimer[20];
+    extern uint16_t interruptTimer[20];
 
 
     void setFrequency (uint32_t freq);
